cloudFunction/analyze/functions/requirements.txt
avro==1.11.3
blinker==1.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chat-downloader==0.2.8
click==8.1.7
cloudevents==1.9.0
colorlog==6.8.2
cryptography==42.0.6
deprecation==2.1.0
docstring_parser==0.16
firebase-admin==6.5.0
firebase-functions==0.4.1
Flask==3.0.3
Flask-Cors==4.0.1
functions-framework==3.5.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
websocket-client==1.8.0
Werkzeug==3.0.2
